server/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
pydantic==2.11.9
python-dotenv>=1.1.1
cryptography>=46.0.1
requests>=2.31.0
httpx==0.25.0
pyjwt==2.10.1